+++ src/autoComplete/AutoCompleteList.unit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow, mount } from 'enzyme'
 
-import AutoCompleteList from './autoCompleteList'
+import AutoCompleteList from './AutoCompleteList'
 import ItemChoice from 'itemChoice'
 
 const fakeItems = [
